requirements.txt
kubernetes
dnspython
ipaddr